ai/talk-with-all-llms/requirements.txt
pydantic-ai-slim[openai]==0.2.16
pydantic-ai-slim[duckduckgo]==0.2.16
